requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0
numpy
scipy
scikit-learn
pandas
Pillow
gdown
